- Replaced all USUL_INC_DIR with CADKIT_ROOT_DIR in project files - Replaced sprintf with GN_FORMAT_STRING macros in GenericNurbs - Small change in Usul/Base/References.h - Added assert type for dot-net environment - Changed unsigned int to size_t in Transpose.h - Made Usul::System::DateTime compile with VC++ 2005 - Made Usul::System::Environment compile with VC++ 2005

+++ Usul/Math/Transpose.h
@@ -35,10 +35,10 @@ template < class Matrix1_, class Matrix2_ > inline void transpose ( const Matrix
   const typename Matrix2_::size_type columns ( ( rows ) ? m.at(0).size() : 0 );
 
   t.resize ( columns );
-  for ( unsigned int c = 0; c < columns; ++c )
+  for ( size_t c = 0; c < columns; ++c )
   {
     t.at(c).resize ( rows );
-    for ( unsigned int r = 0; r < rows; ++r )
+    for ( size_t r = 0; r < rows; ++r )
     {
       if ( columns != m.at(r).size() )
         throw std::runtime_error ( "Error 3987300086: inconsistant sizes in given matrix" );
